Geant4: simplify DD4hep detector construction
