+++ Examples/Algorithms/Geant4/CMakeLists.txt
@@ -1,8 +1,8 @@
 add_library(
   ActsExamplesGeant4 SHARED
+  src/DD4hepDetectorConstruction.cpp
   src/DD4hepToG4Svc.cpp
   src/GeantinoRecording.cpp
-  src/GeoConstruction.cpp
   src/MMDetectorConstruction.cpp
   src/MMEventAction.cpp
   src/MMPrimaryGeneratorAction.cpp
